revamp: minify codes
