--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Controllers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cc8fec8c2f34b4381987360b44c21ae1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
