--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Controllers/RobotActionController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a423db78903874065bc3c6108ba09393
