--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Controllers/RobotPhysicController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ca586183f612a4630a47d6db4cea0f00
