+++ Assets/Dev/Arby/core_actions/Scripts/CoreActions/CoreActionManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f25b0f780301b4d1b8306e7e1154b59b
