+++ Assets/Dev/Arby/core_actions/Scripts/CoreActions/CoreActionRobot.cs
@@ -1,0 +1,25 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace RobotCoreAction
+{
+    public class CoreActionRobot : MonoBehaviour
+    {
+        #region Basic Stats
+        public string Id => System.Guid.NewGuid().ToString();
+        public float MoveSpeed = 4.0f;
+        public float RotateSpeed = 200.0f;
+        #endregion
+
+        #region Dash Stats
+        public float DashSpeed = 5.0f;
+        public float DashDuration = 0.5f;       // Dash duration. 
+        public float DashCooldown = 1.0f;       // Dash cooldown.
+        #endregion
+
+        #region Physics Stats
+        public float StopDelay = 0.5f;           // Time before robot stops.
+        public float SlowDownRate = 2.0f;        // Robot's slowdown rate (velocity and rotation decay). 
+        #endregion
+    }
+}
