--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Input/KeyboardInputProvider.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 958c05a4afb264a369792248fb3b05d8
