--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Input/ScriptInputProvider.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: cad67a9a94b0942c78d1fe506f3b460c
