--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Input/UIInputProvider.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f65ad6a8621394efd96a3a592a27cc95
