--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/RobotStats.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a1ecf39d7f1cf4b3abfecc5827cb1b14
