--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Skills.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4eb77dddaccb54f0a8c31f165442c7bd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
