--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Skills/Skill_Boost.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0e9562ecd0e0a4918b1e327a885e73ad
