--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/Skills/Skill_Stone.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e13929fee07394f77858e7f4a8b11b5a
